Create Badge component
