+++ src/App.jsx
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import { Badge } from './components/common/Badge';
 
 function App() {
   const [count, setCount] = useState(0);
@@ -6,6 +7,7 @@ function App() {
   return (
     <>
       <h1 className='title-24'>Rolling</h1>
+      <Badge />
     </>
   );
 }
