--- src/stores/colorChips.js
@@ -1,6 +1,0 @@
-export const colorChips = [
-  { id: 1, color: 'orange' },
-  { id: 2, color: 'purple' },
-  { id: 3, color: 'blue' },
-  { id: 4, color: 'green' },
-];
